+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.5.0
+==================
+
+* Feature: add send method in Storage to send directly a file without having to prepare the response before
+* Feature: add dispose method in StreamResponse to dispose stream response resources
+
 2.4.0
 ==================